refactor to add everything under chapter1 package
