+++ shino/src/main/java/org/codepolaris/tdd/chapter1/Sum.java
@@ -1,4 +1,4 @@
-package org.codepolaris.tdd;
+package org.codepolaris.tdd.chapter1;
 
 public class Sum implements Expression {
